==> 4_haplogroup_logistic_regression/5_combine_analysis_results.ipynb <==
# Combine ChrY Haplogroup Analysis Results
- **Author(s)** - Frank Grenn
- **Date Started** - April  2021
- **Quick Description** - combine results for AMPPD PD case control, UKBioBank PD case control, UKBioBank PD proxy control, and NEUROX case control results into one table

import pandas as pd

WRKDIR = "$PATH/chrY"

combine_df = pd.DataFrame()

## Major Haplogroup Tests

#amppd pd case control
amp_pd_cc_major = pd.read_csv(f"{WRKDIR}/output_male_hemizygous_only_het_filter_run/haplotype_major_ltrack_case_control_no_gc.csv")
print(amp_pd_cc_major.shape)
print(amp_pd_cc_major.head())

amp_pd_cc_major['cohort'] = 'amp_pd_cc'

amp_pd_cc_major = amp_pd_cc_major[['cohort','haplo_major','controls','control_freq','cases','case_freq','p_logreg','beta_logreg','se_logreg']]
amp_pd_cc_major.columns = ['cohort','haplo_major','controls','control_freq','cases','case_freq','p_logreg','beta_logreg','se_logreg']

combine_df = amp_pd_cc_major.copy()

print(combine_df.shape)
print(combine_df.head())

#ukbb pd case control
ukbb_pd_cc_major = pd.read_csv(f"{WRKDIR}/output_ukbb/haplogroup_major_ltrack_pd_case_control.csv")
print(ukbb_pd_cc_major.shape)
print(ukbb_pd_cc_major.head())

ukbb_pd_cc_major['cohort'] = 'ukbb_pd_cc'
ukbb_pd_cc_major = ukbb_pd_cc_major[['cohort','haplo_major','controls','control_freq','cases','case_freq','p_logreg','beta_logreg','se_logreg']]
ukbb_pd_cc_major.columns = ['cohort','haplo_major','controls','control_freq','cases','case_freq','p_logreg','beta_logreg','se_logreg']
combine_df = combine_df.append(ukbb_pd_cc_major)
print(combine_df.shape)
print(combine_df.head())

#ukbb pd proxy control
ukbb_pd_pc_major = pd.read_csv(f"{WRKDIR}/output_ukbb/haplogroup_major_ltrack_pd_proxy_control.csv")
print(ukbb_pd_pc_major.shape)
print(ukbb_pd_pc_major.head())

ukbb_pd_pc_major['cohort'] = 'ukbb_pd_pc'
ukbb_pd_pc_major = ukbb_pd_pc_major[['cohort','haplo_major','controls','control_freq','proxies','proxy_freq','p_logreg','beta_logreg','se_logreg']]
ukbb_pd_pc_major.columns = ['cohort','haplo_major','controls','control_freq','cases','case_freq','p_logreg','beta_logreg','se_logreg']
combine_df = combine_df.append(ukbb_pd_pc_major)
print(combine_df.shape)
print(combine_df.head())

#neurox pd case control
neurox_pd_cc_major = pd.read_csv(f"{WRKDIR}/output_neurox/haplogroup_major_ltrack_case_control_no_gc.csv")
print(neurox_pd_cc_major.shape)
print(neurox_pd_cc_major.head())

neurox_pd_cc_major['cohort'] = 'neurox_pd_cc'
neurox_pd_cc_major = neurox_pd_cc_major[['cohort','haplo_major','controls','control_freq','cases','case_freq','p_logreg','beta_logreg','se_logreg']]
neurox_pd_cc_major.columns = ['cohort','haplo_major','controls','control_freq','cases','case_freq','p_logreg','beta_logreg','se_logreg']
combine_df = combine_df.append(neurox_pd_cc_major)
print(combine_df.shape)
print(combine_df.head())

combine_df

combine_df.to_csv(f"{WRKDIR}/log_reg_ltrack_results_for_meta.csv",index=None)

### Reformat for Paper

reformat_df = pd.DataFrame()

#amppd pd case control
amp_pd_cc_major = pd.read_csv(f"{WRKDIR}/output_male_hemizygous_only_het_filter_run/haplotype_major_ltrack_case_control_no_gc.csv")
print(amp_pd_cc_major.shape)
print(amp_pd_cc_major.head())


amp_pd_cc_major = amp_pd_cc_major[['haplo_major','controls','control_freq','cases','case_freq','p_logreg','beta_logreg','se_logreg']]
amp_pd_cc_major.columns = ['haplo_major','amp_pd_cc_controls','amp_pd_cc_control_freq','amp_pd_cc_cases','amp_pd_cc_case_freq','amp_pd_cc_p','amp_pd_cc_beta','amp_pd_cc_se']

reformat_df = amp_pd_cc_major.copy()

print(reformat_df.shape)
print(reformat_df.head())

#ukbb pd case control
ukbb_pd_cc_major = pd.read_csv(f"{WRKDIR}/output_ukbb/haplogroup_major_ltrack_pd_case_control.csv")
print(ukbb_pd_cc_major.shape)
print(ukbb_pd_cc_major.head())

ukbb_pd_cc_major= ukbb_pd_cc_major[['haplo_major','controls','control_freq','cases','case_freq','p_logreg','beta_logreg','se_logreg']]
ukbb_pd_cc_major.columns = ['haplo_major','ukbb_pd_cc_controls','ukbb_pd_cc_control_freq','ukbb_pd_cc_cases','ukbb_pd_cc_case_freq','ukbb_pd_cc_p','ukbb_pd_cc_beta','ukbb_pd_cc_se']

reformat_df = pd.merge(left = reformat_df, right = ukbb_pd_cc_major, on = 'haplo_major', how = 'outer')
print(reformat_df.shape)
print(reformat_df.head())

#ukbb pd proxy control
ukbb_pd_pc_major = pd.read_csv(f"{WRKDIR}/output_ukbb/haplogroup_major_ltrack_pd_proxy_control.csv")
print(ukbb_pd_pc_major.shape)
print(ukbb_pd_pc_major.head())


ukbb_pd_pc_major= ukbb_pd_pc_major[['haplo_major','controls','control_freq','proxies','proxy_freq','p_logreg','beta_logreg','se_logreg']]
ukbb_pd_pc_major.columns = ['haplo_major','ukbb_pd_pc_controls','ukbb_pd_pc_control_freq','ukbb_pd_pc_proxies','ukbb_pd_pc_proxy_freq','ukbb_pd_pc_p','ukbb_pd_pc_beta','ukbb_pd_pc_se']

reformat_df = pd.merge(left = reformat_df, right = ukbb_pd_pc_major, on = 'haplo_major', how = 'outer')
print(reformat_df.shape)
print(reformat_df.head())

#neurox pd proxy control
neurox_pd_cc_major = pd.read_csv(f"{WRKDIR}/output_neurox/haplogroup_major_ltrack_case_control_no_gc.csv")
print(neurox_pd_cc_major.shape)
print(neurox_pd_cc_major.head())


neurox_pd_cc_major= neurox_pd_cc_major[['haplo_major','controls','control_freq','cases','case_freq','p_logreg','beta_logreg','se_logreg']]
neurox_pd_cc_major.columns = ['haplo_major','neurox_pd_cc_controls','neurox_pd_cc_control_freq','neurox_pd_cc_cases','neurox_pd_cc_case_freq','neurox_pd_cc_p','neurox_pd_cc_beta','neurox_pd_cc_se']

reformat_df = pd.merge(left = reformat_df, right = neurox_pd_cc_major, on = 'haplo_major', how = 'outer')
print(reformat_df.shape)
print(reformat_df.head())

reformat_df.to_csv(f"{WRKDIR}/log_reg_ltrack_results_for_meta_reformat.csv",index=None)
